Removed the idempotent-ready note, as ready should now be called once.
That is unless you're providing self-created app config instances
to populate (most likely through modify_ or override_settings).
These could also have some custom ready methods, that should be
called despite having previously called ready for another, stored
instance of an app config.

One case to consider is that of providing the very first app config
for a label with a custom ready (for example through INSTALLED_APPS in
settings), after that overriding installed apps using a label of the
app and not getting ready from an apps.py config called (but should
it be?).